Some changes related to a new Clone Dialog
