+++ Mergin/ui/ui_clone_project.ui
@@ -6,32 +6,16 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>362</width>
-    <height>238</height>
+    <width>375</width>
+    <height>130</height>
    </rect>
   </property>
   <property name="windowTitle">
-   <string>Create Mergin Project</string>
+   <string>Clone Mergin Project</string>
   </property>
   <layout class="QVBoxLayout" name="verticalLayout_3">
    <item>
     <layout class="QVBoxLayout" name="verticalLayout_2">
-     <item>
-      <spacer name="verticalSpacer">
-       <property name="orientation">
-        <enum>Qt::Vertical</enum>
-       </property>
-       <property name="sizeType">
-        <enum>QSizePolicy::MinimumExpanding</enum>
-       </property>
-       <property name="sizeHint" stdset="0">
-        <size>
-         <width>20</width>
-         <height>0</height>
-        </size>
-       </property>
-      </spacer>
-     </item>
      <item>
       <layout class="QGridLayout" name="gridLayout">
        <item row="1" column="2">
@@ -49,6 +33,12 @@
        </item>
        <item row="0" column="0">
         <widget class="QLabel" name="projectNamespaceLabel">
+         <property name="sizePolicy">
+          <sizepolicy hsizetype="Expanding" vsizetype="Fixed">
+           <horstretch>0</horstretch>
+           <verstretch>0</verstretch>
+          </sizepolicy>
+         </property>
          <property name="minimumSize">
           <size>
            <width>0</width>
@@ -68,6 +58,12 @@
        </item>
        <item row="0" column="2">
         <widget class="QLabel" name="projectNameLabel">
+         <property name="sizePolicy">
+          <sizepolicy hsizetype="Expanding" vsizetype="Fixed">
+           <horstretch>0</horstretch>
+           <verstretch>0</verstretch>
+          </sizepolicy>
+         </property>
          <property name="minimumSize">
           <size>
            <width>0</width>
@@ -105,57 +101,6 @@
          </property>
          <property name="text">
           <string>/</string>
-         </property>
-        </widget>
-       </item>
-      </layout>
-     </item>
-     <item>
-      <widget class="QCheckBox" name="chk_is_public">
-       <property name="text">
-        <string>Public</string>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <widget class="QRadioButton" name="rad_blank_project">
-       <property name="text">
-        <string>Blank project - Create project on Mergin server</string>
-       </property>
-       <property name="checked">
-        <bool>true</bool>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <widget class="QRadioButton" name="rad_project_dir">
-       <property name="text">
-        <string>Initialize from local drive</string>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <layout class="QHBoxLayout" name="select_dir_layout">
-       <item>
-        <widget class="QLineEdit" name="edit_project_dir">
-         <property name="enabled">
-          <bool>true</bool>
-         </property>
-         <property name="minimumSize">
-          <size>
-           <width>175</width>
-           <height>0</height>
-          </size>
-         </property>
-         <property name="placeholderText">
-          <string>Select project directory</string>
-         </property>
-        </widget>
-       </item>
-       <item>
-        <widget class="QToolButton" name="btn_get_project_dir">
-         <property name="text">
-          <string>...</string>
          </property>
         </widget>
        </item>
